rename BrevModulContext -> useBrevModul og flyttet og renamet testan
