+++ src/frontend/komponenter/Hendelsesoversikt/BrevModul/Brevskjema.tsx
@@ -14,8 +14,8 @@ import BarnBrevetGjelder from './BarnBrevetGjelder';
 import BrevmottakerListe from './BrevmottakerListe';
 import { Brevmal, brevmaler, leggTilValuePåOption, opplysningsdokumenter } from './typer';
 import type { BrevtypeSelect, ISelectOptionMedBrevtekst } from './typer';
+import { useBrevModul } from './useBrevModul';
 import { useBehandling } from '../../../context/behandlingContext/BehandlingContext';
-import { useBrevModul } from '../../../context/BrevModulContext';
 import useDokument from '../../../hooks/useDokument';
 import type { IBehandling } from '../../../typer/behandling';
 import { BehandlingSteg, hentStegNummer } from '../../../typer/behandling';
