+++ src/frontend/komponenter/Hendelsesoversikt/BrevModul/useBrevModul.ts
@@ -5,18 +5,21 @@ import { feil, ok, useFelt, useSkjema, Valideringsstatus } from '@navikt/familie
 import type { Ressurs } from '@navikt/familie-typer';
 import { RessursStatus } from '@navikt/familie-typer';
 
-import { useBehandling } from './behandlingContext/BehandlingContext';
-import type { ISelectOptionMedBrevtekst } from '../komponenter/Hendelsesoversikt/BrevModul/typer';
-import { Brevmal } from '../komponenter/Hendelsesoversikt/BrevModul/typer';
-import { Behandlingstype, BehandlingÅrsak, type IBehandling } from '../typer/behandling';
-import { BehandlingKategori } from '../typer/behandlingstema';
-import type { IManueltBrevRequestPåBehandling } from '../typer/dokument';
-import type { IGrunnlagPerson } from '../typer/person';
-import { PersonType } from '../typer/person';
-import type { IBarnMedOpplysninger } from '../typer/søknad';
-import { Målform } from '../typer/søknad';
-import type { IFritekstFelt } from '../utils/fritekstfelter';
-import { genererIdBasertPåAndreFritekster, lagInitiellFritekst } from '../utils/fritekstfelter';
+import type { ISelectOptionMedBrevtekst } from './typer';
+import { Brevmal } from './typer';
+import { useBehandling } from '../../../context/behandlingContext/BehandlingContext';
+import { Behandlingstype, BehandlingÅrsak, type IBehandling } from '../../../typer/behandling';
+import { BehandlingKategori } from '../../../typer/behandlingstema';
+import type { IManueltBrevRequestPåBehandling } from '../../../typer/dokument';
+import type { IGrunnlagPerson } from '../../../typer/person';
+import { PersonType } from '../../../typer/person';
+import type { IBarnMedOpplysninger } from '../../../typer/søknad';
+import { Målform } from '../../../typer/søknad';
+import type { IFritekstFelt } from '../../../utils/fritekstfelter';
+import {
+    genererIdBasertPåAndreFritekster,
+    lagInitiellFritekst,
+} from '../../../utils/fritekstfelter';
 
 export const hentMuligeBrevmalerImplementering = (
     åpenBehandling: Ressurs<IBehandling>
